pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "milon-exercise-engine"
version = "0.1.0"
description = "AI-powered exercise repetition counter using MediaPipe pose estimation"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
maintainers = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["pose-estimation", "exercise", "fitness", "mediapipe", "computer-vision"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Healthcare Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Multimedia :: Video",
]

dependencies = [
    "mediapipe==0.10.21",
    "opencv-python==4.11.0.86",
    "numpy>=1.26.0",
    "pyyaml==6.0.3",
]

[project.optional-dependencies]
notebook = [
    "jupyter>=1.0",
    "matplotlib>=3.7",
]

[project.urls]
Homepage = "https://github.com/ieee-sb-uniwa/milon-exercise-engine"
Repository = "https://github.com/ieee-sb-uniwa/milon-exercise-engine"

[tool.setuptools.packages.find]
include = ["milon_engine", "milon_engine.*"]

[tool.setuptools.package-data]
milon_engine = ["configs/*.yaml", "configs/*.toml"]

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.ruff]
line-length = 88
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false